add interface template
