+++ sparrow-blog-dao/src/main/java/com/sparrow/blog/dao/PostsDAO.java
@@ -3,7 +3,9 @@ package com.sparrow.blog.dao;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import com.sparrow.blog.dao.dos.PostsEntityDO;
+import org.apache.ibatis.annotations.Mapper;
 
+@Mapper
 public interface PostsDAO extends BaseMapper<PostsEntityDO> {
 
 }
