+++ sparrow-blog-infrastructure/pom.xml
@@ -28,6 +28,11 @@
             <artifactId>sparrow-blog-dao</artifactId>
             <version>1.0.0-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
+            <version>5.2.8.RELEASE</version>
+        </dependency>
     </dependencies>
 
 </project>
