+++ sparrow-blog-main/src/main/java/com/sparrow/blog/boot/Application.java
@@ -21,6 +21,7 @@ import org.springframework.context.ApplicationListener;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.event.ContextClosedEvent;
 import org.springframework.context.event.ContextRefreshedEvent;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
