+++ sparrow-blog-protocol/src/main/java/com/sparrow/blog/protocol/dto/PostsSearchRequestDTO.java
@@ -1,0 +1,4 @@
+package com.sparrow.blog.protocol.dto;
+
+public class PostsSearchRequestDTO {
+}
