+++ sparrow-blog-protocol/src/main/java/com/sparrow/blog/protocol/query/PostsSearchQuery.java
@@ -1,0 +1,4 @@
+package com.sparrow.blog.protocol.query;
+
+public class PostsSearchQuery {
+}
